refactor campaign stats service to model
